Refactoring of cloud-what unit tests
* Split code of unit tests into several files
* Added tearDown methods
